fix: move date-picker vendor files into src
